# apidog-import-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Feedbase Backend API (APIdog Import)",
    "version": "1.0.0",
    "description": "APIdog import file generated from API_FULL_LIST.md."
  },
  "servers": [
    {
      "url": "http://localhost:4560",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/posts/create": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "POST /posts/create",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Post created successfully",
                  "data": {
                    "id": 101
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "postData": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "postType": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "priority": {
                        "type": "integer"
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "postData": {
                  "title": "Add dark mode",
                  "description": "Please add dark mode support for the dashboard.",
                  "postType": "feature_request",
                  "status": "open",
                  "priority": 2
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "POST /posts/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Post retrieved successfully",
                  "data": {
                    "id": 101,
                    "title": "Add dark mode",
                    "description": "Please add dark mode support for the dashboard.",
                    "post_type": "feature_request",
                    "status": "open",
                    "priority": 2,
                    "author_name": "Jane Product",
                    "author_email": "jane@acme.test",
                    "vote_count": 12,
                    "comment_count": 4
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/posts/update/{id}": {
      "put": {
        "tags": [
          "posts"
        ],
        "summary": "PUT /posts/update/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Post updated successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "postData": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "postType": {
                        "type": "string"
                      },
                      "priority": {
                        "type": "integer"
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "postData": {
                  "title": "Add dark mode support",
                  "description": "Please add dark mode and theme persistence.",
                  "postType": "feature_request",
                  "priority": 1
                }
              }
            }
          }
        }
      }
    },
    "/posts/delete/{id}": {
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "DELETE /posts/delete/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Post deleted successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
    "/posts/list": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "POST /posts/list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Posts retrieved successfully",
                  "data": {
                    "posts": [
                      {
                        "id": 101,
                        "title": "Add dark mode",
                        "post_type": "feature_request",
                        "status": "open",
                        "priority": 2,
                        "author_name": "Jane Product",
                        "vote_count": 12
                      }
                    ],
                    "total": 38
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "paginationData": {
                    "type": "object",
                    "properties": {
                      "itemsPerPage": {
                        "type": "integer"
                      },
                      "currentPageNumber": {
                        "type": "integer"
                      },
                      "sortOrder": {
                        "type": "string"
                      },
                      "filterBy": {
                        "type": "string"
                      }
                    }
                  },
                  "filters": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "postType": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "example": {
                "lg": "en",
                "paginationData": {
                  "itemsPerPage": 10,
                  "currentPageNumber": 0,
                  "sortOrder": "desc",
                  "filterBy": ""
                },
                "filters": {
                  "status": "open",
                  "postType": "feature_request"
                }
              }
            }
          }
        }
      }
    },
    "/posts/status/{id}": {
      "patch": {
        "tags": [
          "posts"
        ],
        "summary": "PATCH /posts/status/:id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Post status updated successfully"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  },
                  "newStatus": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en",
                "newStatus": "in_progress"
              }
            }
          }
        }
      }
    },
...
    "/votes/post/{postId}": {
      "post": {
        "tags": [
          "votes"
        ],
        "summary": "POST /votes/post/:postId",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Votes retrieved successfully",
                  "data": [
                    {
                      "id": 301,
                      "post_id": 101,
                      "user_id": 3,
                      "vote_type": "upvote",
                      "full_name": "Jane Product",
                      "email": "jane@acme.test"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
...
    "/comments/post/{postId}": {
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "POST /comments/post/:postId",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "message": "Comments retrieved successfully",
                  "data": [
                    {
                      "id": 401,
                      "post_id": 101,
                      "author_id": 3,
                      "body": "Great request. We need this soon.",
                      "is_edited": 0,
                      "author_name": "Jane Product",
                      "author_email": "jane@acme.test"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lg": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "lg": "en"
              }
            }
          }
        }
      }
    },
...
  }
}